Added Scrubbing feature
Users can now go to the scrub leads webpage.  This page allows users
to upload an excel file.  The user tells the webpage which columns are
firstname, lastname, and a phone number.  The website then scrubs the
excel sheet agaisnt the NoCall list and removes any names it finds.
Helpful for removing leads found in our NoCall list